requirements.txt
apache-airflow
pandas
confluent-kafka
websocket-client
boto3
google-cloud-bigquery
pyarrow
s3fs